+++ addOns/graphql/CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 ### Changed
 - Bug Fixes.
 - Enhanced Support for Script Input Vectors.
+- Options are now exposed through the API.
 
 ## [0.1.0] - 2020-08-28
 - First Version